Externalize socket set to new ublox-sockets crate
