+++ ublox-short-range/src/command/data_mode/mod.rs
@@ -6,6 +6,7 @@ pub mod urc;
 use atat::atat_derive::AtatCmd;
 use responses::*;
 use types::*;
+use ublox_sockets::SocketHandle;
 
 use super::NoResponse;
 
@@ -40,7 +41,7 @@ pub struct ConnectPeer<'a> {
 #[at_cmd("+UDCPC", NoResponse, timeout_ms = 10000)]
 pub struct ClosePeerConnection {
     #[at_arg(position = 0)]
-    pub peer_handle: usize,
+    pub peer_handle: SocketHandle,
 }
 
 /// 5.4 Default remote peer +UDDRP
