+++ ublox-short-range/src/command/ethernet/types.rs
@@ -14,35 +14,35 @@ pub enum OnOff {
 #[repr(u8)]
 pub enum EthernetConfigParameter {
     /// <param_val> decides if the network is active on start up.
-    /// • 0 (default): Inactive
-    /// • 1: active
+    /// - 0 (default): Inactive
+    /// - 1: active
     ActiveOnStartup = 0,
     /// <param_val> Phy support mode
-    /// • 0: disabled
-    /// • 1 (default): enabled
+    /// - 0: disabled
+    /// - 1 (default): enabled
     /// Not available for ODIN-W2 software versions 2.0.0 or 2.0.1. Default PHY values will be used.
     PhySupport = 1,
     /// <param_val> Ethernet speed
-    /// • 0 (default): 100 Mbit/s
-    /// • 1: 10 Mbit/s
+    /// - 0 (default): 100 Mbit/s
+    /// - 1: 10 Mbit/s
     /// Not available for ODIN-W2 software versions 2.0.0 or 2.0.1. Default PHY values will be used.
     Speed = 2,
     /// <param_val> Ethernet Duplex mode
-    /// • 0 (default): Full duplex
-    /// • 1: Half duplex
+    /// - 0 (default): Full duplex
+    /// - 1: Half duplex
     /// Not available for ODIN-W2 software versions 2.0.0 or 2.0.1. Default PHY values will be used.
     DuplexMode = 3,
     /// <param_val> Auto-negotiation (of speed and duplex mode)
-    /// • 0: disabled
-    /// • 1 (default): enabled
+    /// - 0: disabled
+    /// - 1 (default): enabled
     /// Not available for ODIN-W2 software versions 2.0.0 or 2.0.1. Default PHY values will be used.
     AutoNegotiation = 4,
     /// <param_val> is the Phy address. The factory default value is 0x3 (for ODIN) and 0x0
     /// (for NINA-W13 and NINA-W15).
     PhyAddress = 5,
     /// IPv4 Mode - <param_val1> to set the way to retrieve an IP address
-    /// • 1 (default): Static
-    /// • 2: DHCP
+    /// - 1 (default): Static
+    /// - 2: DHCP
     IPv4Mode = 100,
     /// <param_val> is the IPv4 address. The factory default value is 0.0.0.0
     IPv4Address = 101,
@@ -58,39 +58,39 @@ pub enum EthernetConfigParameter {
     SecondaryDNS = 105,
     /// Address conflict detection. The factory default value is 0 (disabled). This tag is
     /// supported by ODIN-W2 from software version 6.0.0 onwards only.
-    /// • 0: Disabled
-    /// • 1: Enabled
+    /// - 0: Disabled
+    /// - 1: Enabled
     AddressConflictDetection = 107,
 }
 
 #[derive(Clone, PartialEq, AtatEnum)]
 pub enum EthernetConfig {
     /// <param_val> decides if the network is active on start up.
-    /// • 0 (default): Inactive
-    /// • 1: active
+    /// - 0 (default): Inactive
+    /// - 1: active
     #[at_arg(value = 0)]
     ActiveOnStartup(OnOff),
     /// <param_val> Phy support mode
-    /// • 0: disabled
-    /// • 1 (default): enabled
+    /// - 0: disabled
+    /// - 1 (default): enabled
     /// Not available for ODIN-W2 software versions 2.0.0 or 2.0.1. Default PHY values will be used.
     #[at_arg(value = 1)]
     PhySupport(OnOff),
     /// <param_val> Ethernet speed
-    /// • 0 (default): 100 Mbit/s
-    /// • 1: 10 Mbit/s
+    /// - 0 (default): 100 Mbit/s
+    /// - 1: 10 Mbit/s
     /// Not available for ODIN-W2 software versions 2.0.0 or 2.0.1. Default PHY values will be used.
     #[at_arg(value = 2)]
     Speed(EthernetSpeed),
     /// <param_val> Ethernet Duplex mode
-    /// • 0 (default): Full duplex
-    /// • 1: Half duplex
+    /// - 0 (default): Full duplex
+    /// - 1: Half duplex
     /// Not available for ODIN-W2 software versions 2.0.0 or 2.0.1. Default PHY values will be used.
     #[at_arg(value = 3)]
     DuplexMode(EthernetDuplexMode),
     /// <param_val> Auto-negotiation (of speed and duplex mode)
-    /// • 0: disabled
-    /// • 1 (default): enabled
+    /// - 0: disabled
+    /// - 1 (default): enabled
     /// Not available for ODIN-W2 software versions 2.0.0 or 2.0.1. Default PHY values will be used.
     #[at_arg(value = 4)]
     AutoNegotiation(OnOff),
@@ -99,8 +99,8 @@ pub enum EthernetConfig {
     #[at_arg(value = 5)]
     PhyAddress(u32),
     /// IPv4 Mode - <param_val1> to set the way to retrieve an IP address
-    /// • 1 (default): Static
-    /// • 2: DHCP
+    /// - 1 (default): Static
+    /// - 2: DHCP
     #[at_arg(value = 100)]
     IPv4Mode(IPv4Mode),
     /// <param_val> is the IPv4 address. The factory default value is 0.0.0.0
@@ -122,8 +122,8 @@ pub enum EthernetConfig {
     SecondaryDNS(#[at_arg(len = 16)] Ipv4Addr),
     /// Address conflict detection. The factory default value is 0 (disabled). This tag is
     /// supported by ODIN-W2 from software version 6.0.0 onwards only.
-    /// • 0: Disabled
-    /// • 1: Enabled
+    /// - 0: Disabled
+    /// - 1: Enabled
     #[at_arg(value = 107)]
     AddressConflictDetection(OnOff),
 }
@@ -131,31 +131,31 @@ pub enum EthernetConfig {
 #[derive(Clone, PartialEq, AtatEnum)]
 pub enum EthernetConfigR {
     /// <param_val> decides if the network is active on start up.
-    /// • 0 (default): Inactive
-    /// • 1: active
+    /// - 0 (default): Inactive
+    /// - 1: active
     #[at_arg(value = 0)]
     ActiveOnStartup(OnOff),
     /// <param_val> Phy support mode
-    /// • 0: disabled
-    /// • 1 (default): enabled
+    /// - 0: disabled
+    /// - 1 (default): enabled
     /// Not available for ODIN-W2 software versions 2.0.0 or 2.0.1. Default PHY values will be used.
     #[at_arg(value = 1)]
     PhySupport(OnOff),
     /// <param_val> Ethernet speed
-    /// • 0 (default): 100 Mbit/s
-    /// • 1: 10 Mbit/s
+    /// - 0 (default): 100 Mbit/s
+    /// - 1: 10 Mbit/s
     /// Not available for ODIN-W2 software versions 2.0.0 or 2.0.1. Default PHY values will be used.
     #[at_arg(value = 2)]
     Speed(EthernetSpeed),
     /// <param_val> Ethernet Duplex mode
-    /// • 0 (default): Full duplex
-    /// • 1: Half duplex
+    /// - 0 (default): Full duplex
+    /// - 1: Half duplex
     /// Not available for ODIN-W2 software versions 2.0.0 or 2.0.1. Default PHY values will be used.
     #[at_arg(value = 3)]
     DuplexMode(EthernetDuplexMode),
     /// <param_val> Auto-negotiation (of speed and duplex mode)
-    /// • 0: disabled
-    /// • 1 (default): enabled
+    /// - 0: disabled
+    /// - 1 (default): enabled
     /// Not available for ODIN-W2 software versions 2.0.0 or 2.0.1. Default PHY values will be used.
     #[at_arg(value = 4)]
     AutoNegotiation(OnOff),
@@ -164,8 +164,8 @@ pub enum EthernetConfigR {
     #[at_arg(value = 5)]
     PhyAddress(u32),
     /// IPv4 Mode - <param_val1> to set the way to retrieve an IP address
-    /// • 1 (default): Static
-    /// • 2: DHCP
+    /// - 1 (default): Static
+    /// - 2: DHCP
     #[at_arg(value = 100)]
     IPv4Mode(IPv4Mode),
     /// <param_val> is the IPv4 address. The factory default value is 0.0.0.0
@@ -187,8 +187,8 @@ pub enum EthernetConfigR {
     SecondaryDNS(#[at_arg(len = 16)] Ipv4Addr),
     /// Address conflict detection. The factory default value is 0 (disabled). This tag is
     /// supported by ODIN-W2 from software version 6.0.0 onwards only.
-    /// • 0: Disabled
-    /// • 1: Enabled
+    /// - 0: Disabled
+    /// - 1: Enabled
     #[at_arg(value = 107)]
     AddressConflictDetection(OnOff),
 }
