--- ublox-short-range/src/socket/meta.rs
@@ -1,13 +1,0 @@
-use super::{ChannelId, SocketHandle};
-
-/// Network socket metadata.
-///
-/// This includes things that only external (to the socket, that is) code
-/// is interested in, but which are more conveniently stored inside the socket itself.
-#[derive(Debug, Default)]
-pub struct Meta {
-    /// Handle of this socket within its enclosing `SocketSet`.
-    /// Mainly useful for debug output.
-    pub(crate) handle: SocketHandle,
-    pub(crate) channel_id: ChannelId,
-}
